requirements.txt
anthropic
python-dotenv
# system install portaudio
pyaudio
speechrecognition
elevenlabs
pydub
groq
pillow